Add openmpi support in mpich2-yarn.
Mpich2 support -launch ssh and -hosts a:1 b:2 to launch mpi apps but openmpi does not.
When we try to launch openmpi apps, we need to first prepare a machine file to show the
process launch metrics. Currently, we just directly launch all the apps in openmpi format.
This needs to be fixed in the next.